Bug fix: composite payload httpHeaders should be map[string]string instead of map[string][]string
Fixes Issue #56

A drawback of this fix is, if the user passes in multiple values for a key, Salesforce would still return an error.

Another possible fix is to change the type of Subrequester.HTTPHeaders from http.Header to map[string]string but we probably don't want to do that because callers may expect to be able to pass in a http.Header type for that field.